ContactFrom_v15/requirements.txt
pandas==0.24.2
Flask==0.12.2
matplotlib==3.1.0
numpy==1.16.3
nltk==3.4.5
smooch==2.9.0
PyJWT==1.7.1
wordcloud==1.6.0
